+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Fix #1351 tabs is-disabled css rule
 * Fix #1352 numberinput when empty string
 * Fix #1382 numberinput layout
+* Fix #1390 missing prop to dialog type
 * Fix #1402 radio button on IE11 (thanks @timEulitz)
 
 ## 0.7.5